Split metadata tag entry layout
