+++ src/gui_app/folder_browser.rs
@@ -323,6 +323,7 @@ pub(super) use types::{
 
 mod tag_completion;
 mod tag_editor;
+mod tag_entry_layout;
 mod view;
 pub(super) use view::folder_browser_view;
